Changed actions into thunks
